impl Lox.runPrompt()
extends the "run" facility in java.mak to allow redirection
of output via "tee" so it remains visible, and shows the
redirections on the "JAVA RUN" log line.

adds rules to j/jlox/GNUmakefile to support running the
interpreter on a collection of input files. There is no
facility (yet) to verify correctness of the result.

start with a test case that just confirms that the input
text is seen by the interpreter.