add //-comments to Scanner.scanToken()
IMPL Scanner.peek()

Made the 2nd line of repl02.in a comment.
Note that the scanner records only an EOF token for this line.

Also note that args-1.out reflects the fact that the first
line of hello.lex is now scanned as a comment (a lot of the
Unexpected Character errors go away).