ADD whitespace to Scanner.scanToken
includes updating the line number, which adjusts all of
the errors in the test logs that are on lines other than 1.

I embedded a \r in the middle of ta line in repl02.in. If it
were at the end of the line, it would be too easy for "helpful"
code to notice the CRLF and "fix" it.

repl03 is text presented int he book with the comment "It can
handle fairly free-form code like:" ...