ADD /*...*/ comments to Scanner.scanToken()
Test inputs added to hello.lox (which includes a
multi-line comment) and repl04.in ... note that
the REPL processes one line at a time, so a newline
in a comment turns into an unterminated comment.

Since the only changes to hello.lox are comments,
the output file (args-1.out) does not change.