impl GenerateAst.defineAst()
initially, just write the file that creates the
outer abstract class.

Track the generated file in Git so we can monitor
how it changes from commit to commit.

So the rule from now on is to "make world" before
committing, so the generated source is present.

Also, a "make world" is suggested immediately after
initial clone or changing branches, to avoid the case
where the generated file up to date but for the wrong
version of GenerateAst.