ADD GenerateAst accept() code gen
Net effect on Expr.java:

- base accept(Visitor<R>) method in the abstract class
- overide accept() methods in each of the static classes

Note: we can now build self-contained Visitors that provide
some code for each kind of Expr, without having to come back
to Expr to add anything about the new Visitor.

This means that if we add new kinds of Expr, we would then
need to go back and fix up every Visitor. That's the trade-off.