Generate the "Stmt" AST class
- extend GenerateAst to include Stmt and its concrete classes
- tell the build system we now build Stmt.java
- capture the initial Stmt.java in Git

notice that "make tests" was not assuring the build had
been done. Add "make build" that does the build (incluiding
assuring GenerateAst has been run), and rework other targets
to depend on the "build" target.

reflect this in j/java.mak as well

finally, fix whitespace in the makefiles
- use tabs for indentation
- do not use tabs for alignment

resisting the temptation to divide this commit into parts.