FIX missing build dependencies
For example, if we add a class to Expr.java, we need to
rebuilt AstPrinter and Interpreter which have Visitors
that need to exist for every Expr class.

Also, runtime of javac is dominated by the cost to start
up the JVM, so we want to run javac just once for all of
the files in a package.

Currently there is no significant time savings by leaving
out the .java files that we "know" do not need to be
rebuilt, so if any class file in a package needs to be
rebuilt, we just rebuild them all.