Opinions and Build System Tweaks
I have strong opinions about how C99 code should be written,
so read the Opinions.org file for more. I will try to keep
the committed sources close to the text of the book. Some
changes are just relaxing from the constraints of what source
has to look like when publishing to print, others may be from
strongly held opinions about what is Right.

I use GNU "indent" (curently version 2.2.12) to adjust the
style of my C source code. The file c/clox/.indent.pro will
encode the style of the code in this project. Note that it
lists the intentional ways the style of this project differs
from the stock GNU style that is the "indent" default.