Build System for clox, and tweaks
This project diverged from my normal $Q, $C, and $I macros.
With this commit I return to what I consider normal. See the
comments in GNUmakefile for their definitions.

In light of the above, added comments for more of the macros
defined in the top-level GNUmakefile as I checked them for
consistency with my ancient master Projects area.

Drop the support for my "ru" monitor. This just adds complexity
to the build without adding value.

Fix the "compile .c to .o" rule. "$^" was printing the relative
paths to the source file and, when a .d file was present, the
paths to every header file it depended on. Ugh. Instead, just
print the basename of the source file.

Compilation and Linking is $Q, not $C, because we want the build
to stop before building anything depending on the results. Note
that j/java.mak has a similar adjustment.

add c/clox/GNUmakefile which introduces the "indent" target,
currently only run on explicit command. Oh, and we ask the build
to use the "c99" front-end script for GCC that constrains it to
using C99, not some newer version of the language.