import "format" from live-coding project
Time to enforce header include order. I got this working
back in the live-coding project but did not push it all
the way back to my master Projects area, so go grab it.

Adding "main.h" for consistency and to exercise logic.

I suspect "main.h" might come in handy in a small number
of potential projects, so I'm not averse to keeping it
even if it is only ever this one redundant declaration.