Combine indent with fixing headers
had a situation where the header fix made a change
that indent removed, causing that file to always
be updated by a make format.

oh, and now we no longer have blank lines at the
top of C files that have no H file.

Discarding main.h, we can put it back if it turns
out to be actually useful.

Starting to think in terms of YAGNI.