add disassembleChunk and friends
Note that the chunk BIST now outputs a disassembly
of its test chunk to stdout.

I can't help but think that the disassembly would be
more useful if it returned a string that could be
examined by the test. Perhaps in the future but this
is not important enough to motivate me to upend the
debug code that will likely be short-lived.