refine error reporting a smidge
fix the build so nonzero exit from bist is a build error.

redirect errors from grow_checked_array (in memory.c)
to use assert rather than directly aborting.

fix up wording for assertion failure messages.

skip printing file and line if file is null or line is zero.

Assure everything printed to stdout is flushed before
the formatted message goes to stderr.

Add a newline to stdout because otherwise the message
will not be at the start of the output line when stdout
and stderr go to the same destination.

Add a blank line after the message so our formatted
message does not get mushed into the error from make
that the program had nonzero exit status.