26.5.1 Weak references and the string pool
NOTE: due to GC bugs that remain, this code will be removing
entries from the string table that are still needed, resulting
in (for example) a lot of "undefined variable" errors.

As a consequence, I have commented out the line that removes
the entries from the table and left a "TODO" there to reenable
the line when we kill the known GC bugs.