26.7.1 Fix GC bug in addConstant
First effect is that the item being protected will now be
seen first in the mark list, and not seen later; it will
then be seen at the end of the blacken list. These cases
are where the item did not need protection.

HOWEVER, we also see cases (see ch25-03) where we are now
marking and blackening new things. These are cases where
the items newly apparing in the logs would have been swept
away while being created.