tool edit: difflinks
sometimes the "make" job output will have "diff" lines indiating
the addition or deletion of messages that would be parsable by
the editor, but with a leading "+" or "-" ... make some symlinks
allowing the editor to locate the indicated file.

run this script to create +src -src +inc -inc when needed.